+++ src/nanonext.h
@@ -26,6 +26,7 @@
 #ifdef NANONEXT_PROTOCOLS
 #include <nng/protocol/bus0/bus.h>
 #include <nng/protocol/pair0/pair.h>
+#include <nng/protocol/pair1/pair.h>
 #include <nng/protocol/pubsub0/pub.h>
 #include <nng/protocol/pubsub0/sub.h>
 #include <nng/protocol/pipeline0/pull.h>